requirements.txt
alembic==1.0.8
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-HTTPAuth==3.2.4
flask-marshmallow==0.10.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==2.19.1
marshmallow-sqlalchemy==0.16.1
numpy==1.16.2
Pillow==5.4.1
python-dateutil==2.8.0
python-editor==1.0.4
requests==2.21.0
scipy==1.2.1
sightengine==1.4.0
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1